requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
curl_cffi==0.12.0
discord-protos==0.0.2
discord.py-self==2.0.1
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
pip==25.1.1
propcache==0.3.2
protobuf==5.26.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.1
tzdata==2025.2
tzlocal==5.3.1
yarl==1.20.1
flask 